[AIRFLOW-6721] Move HdfsSensor test to separate module
Move the HdfsSensor test in the operators module into a new test
module and use the shared test environment.
